fix: refactor handlers to own files
